### 2019-04-14 #1380 @whedon
[PRE REVIEW]: coronagraph: Telescope Noise Modeling for Exoplanets in Python
**Submitting author:** @jlustigy (<a href="http://orcid.org/0000-0002-0746-1980">Jacob Lustig-Yaeger</a>)
**Repository:** <a href="https://github.com/jlustigy/coronagraph" target ="_blank">https://github.com/jlustigy/coronagraph</a>
**Version:** v0.1
