### 2019-04-15 #1380 @labarba
@arfon, I think this submission is right up your alley. Do you want to take it as handling editor?
