+++ scripts/ground_demo-1.py
@@ -1,0 +1,4 @@
+from scripts import ground_demo
+from coronagraph import plot_setup
+plot_setup.setup()
+ground_demo._test()
