+++ scripts/transit_demo-1.py
@@ -1,0 +1,4 @@
+from scripts import transit_demo
+from coronagraph import plot_setup
+plot_setup.setup()
+transit_demo.earth_analog_transits()
