+++ scripts/luvoir_demo-1.py
@@ -1,0 +1,4 @@
+from scripts import luvoir_demo
+from coronagraph import plot_setup
+plot_setup.setup()
+luvoir_demo._test()
