+++ _sources/examples.rst.txt
@@ -1,0 +1,10 @@
+Examples
+========
+
+A collection of Jupyter Notebooks that demonstrate how to use :py:obj:`coronagraph`.
+
+.. toctree::
+   :maxdepth: 1
+
+   Degrading a Spectrum <notebooks/degrading_a_spectrum.ipynb>
+   Secondary Eclipse Spectroscopy <notebooks/eclipse_example.ipynb>
