+++ _sources/scripts/luvoir_demo.rst.txt
@@ -1,0 +1,2 @@
+.. automodule:: scripts.luvoir_demo
+  :members:
