+++ _sources/install.rst.txt
@@ -1,0 +1,16 @@
+Installation
+============
+
+The simplest option is to install using :py:obj:`pip`:
+
+.. code-block:: bash
+
+   pip install coronagraph
+
+Users may also clone the repository on Github:
+
+.. code-block:: bash
+
+   git clone https://github.com/jlustigy/coronagraph.git
+   cd coronagraph
+   python setup.py
