+++ _sources/api/degrade_spec.rst.txt
@@ -1,0 +1,6 @@
+Degrade Spectra
+===============
+
+.. automodule:: coronagraph.degrade_spec
+   :members:
+   :undoc-members:
