### 2019-04-19 17dd4a4
Changed doc links to gh-pages to fix stretched plots
