### 2019-04-23 0bce723
modifications for planet-star separation dependent coronagraph throughput and contrast (optionally)
