### 2019-05-29 #1387 @jlustigy
👋 @arfon and @danielskatz, for what it's worth, I think @lpueyo would make a great reviewer for this paper/code and I'm willing to wait until he has the time to complete it. But I also understand if JOSS policy favors an alternate reviewer.
