### 2019-06-10 #1387 @lpueyo
This is very useful software software that greatly needed to me made public. I am glad that the authors took this initiative. I have ran all example notebooks and changed parameters. The results make sense. I did not find any major software issues. I did notice that the reference file “earth_avg_hitran2012_300_100000cm.trns” is missing and preventing one of the notebooks to completely run.  I would suggest the authors to double check for the presence of all the reference files.

While the software in itself will be useful to our community, I do have a few  outstanding  issues with the documentation:

-	I was not able to find a detailed API description. This is needed to make the software more broadly accessible.

-	The statement of purpose in vague in the actual paper. Why do we need this code? Who should use this? Are there caveats for which the results ought to be taken with a grain of salt?

- Also, there are already a few public coronagraph simulation codes, say for instance https://github.com/wfirst-cgi or https://github.com/dsavransky/EXOSIMS or the python version of the code in http://spie.org/Publications/Proceedings/Paper/10.1117/12.2274396. Since I have been involved (sometime remotely) with most of these projects, it is very clear to me when to you use which package. However, it is not clear to any random user. The paper ought to explain this. It does not have to be very detailed, but the code associated with this submission is the only one that features the “reverse” calculation (SNR to exposure time) so this needs to be highlighted. A

-	An example notebook comparing the submitted code with one of these already existing in the context of the “forward” calculation would be nice, but it might be too much work for little benefit to the users.
