### 2019-06-27 d0bb223
doc updates and data_files added to setup.py
