+++ docs/api/count_rates.rst
@@ -1,15 +1,6 @@
 Coronagraph Noise Modeling
 ==========================
 
-The crux of ``coronagraph`` noise modeling is determining the photon count rate
-incident upon the detector due to the planet and an assortment of
-different noise sources. The following classes and functions serve as your
-interface to the photon count rate calculations.
-
 .. automodule:: coronagraph.count_rates
    :members:
    :undoc-members:
-
-.. automodule:: coronagraph.count_rates_wrapper
-  :members:
-  :undoc-members:

[…]
