### 2019-06-28 c846c8f
one more file added to setup.py data_files; more doc updates
