+++ setup.py
@@ -49,6 +49,7 @@ setup(name='coronagraph',
                   "coronagraph/planets/EarlyMars_geo_albedo.txt",
                   "coronagraph/planets/EarlyVenus_geo_albedo.txt",
                   "coronagraph/planets/earth_avg_hitran2012_300_100000cm.trnst",
+                  "coronagraph/planets/earth_avg_hitran2012_300_100000cm_toa.rad", 
                   "coronagraph/planets/Hazy_ArcheanEarth_geo_albedo.txt",
                   "coronagraph/planets/Jupiter_geo_albedo.txt",
                   "coronagraph/planets/Mars_geo_albedo.txt",
