### 2019-06-29 573a616
more doc updates
