+++ docs/install.rst
@@ -7,7 +7,7 @@ The simplest option is to install using :py:obj:`pip`:
 
    pip install coronagraph
 
-Users may also clone the repository on Github:
+Users may also clone the repository on Github for the latest version of the code:
 
 .. code-block:: bash
 
