### 2019-06-29 e41f49a
possible travis xvfb fix
