+++ .travis.yml
@@ -36,8 +36,11 @@ install:
 
 before_script: # configure a headless display to test plot generation
   - "export DISPLAY=:99.0"
-  - "sh -e /etc/init.d/xvfb start"
-  - sleep 3 # give xvfb some time to start
+#  - "sh -e /etc/init.d/xvfb start"
+#  - sleep 3 # give xvfb some time to start
+
+services:
+  - xvfb
 
 script:
   #- mkdir -p $TEST_DIR
