### 2019-06-30 #1387 @arfon
The author has responded to say that he's made a number of improvements to his submission but still needs a little more time to finish them up. He suggested he would be able to complete these in the next two weeks.
