### 2019-07-14 4294d38
JOSS paper updates for review
