+++ docs/paper/paper.bib
@@ -131,3 +131,17 @@ archivePrefix = "arXiv",
   url = {http://luvoir.stsci.edu/coron_model},
   urldate = {2018-10-13}
 }
+
+@MISC{exosims_code,
+       author = {{Savransky}, Dmitry and {Delacroix}, Christian and {Garrett}, Daniel},
+        title = "{EXOSIMS: Exoplanet Open-Source Imaging Mission Simulator}",
+     keywords = {Software, NASA},
+         year = "2017",
+        month = "Jun",
+          eid = {ascl:1706.010},
+        pages = {ascl:1706.010},
+archivePrefix = {ascl},
+       eprint = {1706.010},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2017ascl.soft06010S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
