+++ docs/paper/paper.md
@@ -38,12 +38,29 @@ Briefly, ``coronagraph`` uses analytic relations to calculate the photon count
 rates for a variety of astrophysical, telescope, and
 instrumental noise sources. These include photons from
 coronagraph speckles, zodiacal and exozodiacal dust, telescope thermal emission,
-dark current, and read noise. The model also includes Earth's atmospheric
+dark current, read noise, and clock-induced charge.
+The model also includes Earth's atmospheric
 transmission and emission spectrum from the UV through the near IR for
 ground-based telescope modeling [@Meadows2018]. Photons from a user-provided
 exoplanet source spectrum are compared against sources of noise to provide
-signal-to-noise ratios and synthetic observations that may be used to assess
-the exoplanet science capabilities of near- and far-term future telescopes.  
+wavelength-dependent signal-to-noise (S/N) ratios and synthetic observations.
+The model also features "reverse" S/N calculations, which provide the requisite
+exposure times per wavelength interval such that a user-specified S/N is
+achieved on the exoplanet spectrum.
+
+The ``coronagraph`` model was designed to assess exoplanet science capabilities
+for near- and far-term future telescopes. Since the computations are all
+relatively simple and analytic, ``coronagraph`` is ideal for fast sweeps across
+a range of possible telescope, instrumental, astrophysical, and planetary
+parameters to investigate primary noise sources and the feasibility of different
+science case. However, ``coronagraph`` may be impractical for existing
+telescopes or telescope concepts that have mature designs due to the
+"first order" nature of the model.
+Additionally, systematic errors (e.g. from speckle subtraction), stray light
+from companion stars and/or background objects, and time-dependent (red) noise
+terms are not considered. More specific or detailed use cases than those
+outlined above may be appropriate for the Exoplanet Open-Source Imaging Mission
+Simulator (EXOSIMS) [@exosims_code].
 
 ``coronagraph`` has already been used in numerous practical science
 applications. This includes peer-reviewed work on the potential for
@@ -62,7 +79,11 @@ relevant to a far-infrared surveyor mission concept [@Cooray2017].
 
 # Acknowledgements
 
-This work was supported by the NASA Astrobiology Institute's Virtual Planetary
-Laboratory under Cooperative Agreement number NNA13AA93A.
+This work was performed as part of NASA's Virtual Planetary Laboratory,
+supported by the National Aeronautics and Space Administration through the NASA
+Astrobiology Institute under solicitation NNH12ZDA002C and Cooperative Agreement
+Number NNA13AA93A, and by the NASA Astrobiology Program under grant
+80NSSC18K0829 as part of the Nexus for Exoplanet System Science (NExSS)
+research coordination network.
 
 # References
