### 2019-07-14 #1387 @jlustigy
Thank you @lpueyo for the review (especially during your paternity leave!). My apologies for the delay, it has been a busy summer for me so far.

>  I did notice that the reference file “earth_avg_hitran2012_300_100000cm.trns” is missing and preventing one of the notebooks to completely run. I would suggest the authors to double check for the presence of all the reference files.

I have updated the `setup.py` file to include `data_files` including the file that you noted was missing. However, I have not yet updated the package version number or version on PyPI, because I wanted to double check with @arfon the correct procedure for doing so since this review is tied to a specific version number of my code. I would appreciate guidance on this if possible! In the meantime, @lpueyo, if you clone the github repository you should find that the code now detects the file: “earth_avg_hitran2012_300_100000cm.trnst”.

> I was not able to find a detailed API description. This is needed to make the software more broadly accessible.

Regarding the API, I have added significantly more detail to the existing API, which can be found within the code documentation here: https://jlustigy.github.io/coronagraph/api.html.

> The statement of purpose in vague in the actual paper. Why do we need this code? Who should use this? Are there caveats for which the results ought to be taken with a grain of salt?

The paper (paper.md) was modified in the following ways:
* added a statement about the code allowing for "reverse" S/N calculations (requisite exposure time to get a given S/N);
* added a paragraph describing the general purpose of the code, including when it may be useful and under what circumstances it may not be the code of choice;
* updated the Acknowledgements to reflect newer funding sources
* clock-induced photon charge is now included as a noise source;

Regrading the other similar codes and suggested example, I verified (offline) that this code (`coronagraph`) agrees with the model originally written in IDL by Tyler Robinson, which was published in https://iopscience.iop.org/article/10.1088/1538-3873/128/960/025003. Note that the WFIRST CGI coronagraph model (https://github.com/wfirst-cgi) is essentially the same (original IDL) code, albeit with parameters hard coded for the WFIRST mission. I very little experience using the EXOSIMS code, and although I really like the idea of adding an example notebook to my docs comparing it to my code, this may take a significant amount of time for me to create an apples-to-apples benchmarking test. Unless acceptance of this model in JOSS hinges on this, it may be appropriate for me to add this as an “enhancement" issue to my github repo, so I can complete it at a later time.

Thank you all for your patience awaiting my updates on this!
