### 2019-08-22 #1387 @lpueyo
Excellent job @jlustigy the new API helps a lot. Thank you for doing it. I am happy about this submission and looking forward to using it for my research! This is closed as far as I am concerned.
