### 2019-08-22 #1387 @arfon
Also, please merge this PR @jlustigy - it should fix a few formatting issues with your paper: https://github.com/jlustigy/coronagraph/pull/3
