### 2019-08-23 bdf9f78
bumped version to 1.0 following completion of JOSS review
