### 2019-08-23 #1387 @jlustigy
Thanks again @lpueyo for the helpful review!

@arfon - I released the updated code as v1.0 and archived it on Zenodo [here](https://zenodo.org/record/3375103). The DOI for the archive is: 10.5281/zenodo.3375103. Thank you for all the help during the review process!
