### 2019-08-23 #1387 @jlustigy
@arfon - I was just notified of two minor typos associated with one of the coauthor's name and affiliation. I just pushed the corrections to [coronagraph master](https://github.com/jlustigy/coronagraph/tree/master). Is it possible to recompile the paper at this point?
