### 2019-08-23 #1387 @arfon
> @arfon - I was just notified of two minor typos associated with one of the coauthor's name and affiliation. I just pushed the corrections to coronagraph master. Is it possible to recompile the paper at this point?

Sure thing. I've updated the paper now. Note, it can take a while to show updated on the JOSS site as there's caching in place for the PDFs.
