### 2019-08-23 #1387 @jlustigy
Excellent! I also updated the Zenodo archive to match the name change. Thanks for this! JOSS for the win 🎉
